# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ElSpa Payroll System API",
    "description": "Complete RESTful API for ElSpa payroll management system.\n\nFeatures:\n- Employee management (CRUD)\n- Cash Advance (CA) tracking and settlement\n- Attendance logging with OT/late calculations\n- Philippine holiday management\n- Payroll period management\n- Automated payroll calculation with detailed breakdown\n- Payroll approval workflow\n\n**Authentication:** JWT Bearer tokens (required for all endpoints in production)\n\n**Error Handling:** All endpoints return appropriate HTTP status codes with error messages\n\n**Pagination:** List endpoints support skip/limit parameters\n",
    "version": "1.0.0",
    "contact": {
      "name": "ElSpa Development",
      "email": "support@elspa.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.elspa.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
    "/api/payroll/employees": {
      "post": {
        "tags": [
          "Employees"
        ],
        "summary": "Create Employee",
        "description": "Create a new employee record in the payroll system.\n\nEmployee types:\n- therapist: Service therapist (can have commission)\n- nail: Nail technician (can have commission)\n- driver: Transportation driver (eligible for meal allowance)\n- maintenance: Maintenance staff\n- hollys: Hollys Coffee staff\n- manager: Manager/admin\n\nPay groups:\n- weekly: Weekly payroll cycle\n- biweekly: Bi-weekly payroll cycle\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeCreate"
              },
              "examples": {
                "therapist": {
                  "summary": "Create Therapist",
                  "value": {
                    "name": "Maria Santos",
                    "phone": "0917-123-4567",
                    "employee_type": "therapist",
                    "pay_group": "weekly",
                    "base_salary": 8000.0,
                    "commission_rate": 10.0,
                    "hire_date": "2025-01-15"
                  }
                },
                "driver": {
                  "summary": "Create Driver",
                  "value": {
                    "name": "Juan Cruz",
                    "phone": "0917-234-5678",
                    "employee_type": "driver",
                    "pay_group": "biweekly",
                    "base_salary": 10000.0,
                    "commission_rate": 0,
                    "hire_date": "2024-06-01"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Employee created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponse"
                },
                "example": {
                  "id": 1,
                  "name": "Maria Santos",
                  "phone": "0917-123-4567",
                  "employee_type": "therapist",
                  "pay_group": "weekly",
                  "base_salary": 8000.0,
                  "commission_rate": 10.0,
                  "hire_date": "2025-01-15",
                  "is_active": true,
                  "created_at": "2026-05-21T10:30:00",
                  "updated_at": "2026-05-21T10:30:00"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (validation error)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "base_salary must be non-negative"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "List Employees",
        "description": "Retrieve paginated list of employees with optional filtering by type",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Number of records to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 100
            },
            "description": "Maximum number of records to return"
          },
          {
            "name": "employee_type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "therapist",
                "nail",
                "driver",
                "maintenance",
                "hollys",
                "manager"
              ]
            },
            "description": "Filter by employee type"
          }
        ],
        "responses": {
          "200": {
            "description": "List of employees",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployeeResponse"
                  }
                },
                "example": [
                  {
                    "id": 1,
                    "name": "Maria Santos",
                    "phone": "0917-123-4567",
                    "employee_type": "therapist",
                    "pay_group": "weekly",
                    "base_salary": 8000.0,
                    "commission_rate": 10.0,
                    "hire_date": "2025-01-15",
                    "is_active": true,
                    "created_at": "2026-05-21T10:30:00",
                    "updated_at": "2026-05-21T10:30:00"
                  }
                ]
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/payroll/employees/{employee_id}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "Get Employee",
        "description": "Retrieve detailed information for a specific employee",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Employee ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Employee details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "\uc9c1\uc6d0\uc744 \ucc3e\uc744 \uc218 \uc5c6\uc2b5\ub2c8\ub2e4"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Employees"
        ],
        "summary": "Update Employee",
        "description": "Update employee information",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Employee ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Employees"
        ],
        "summary": "Delete Employee",
        "description": "Soft delete an employee (marks is_active as false)",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Employee ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Employee deleted successfully"
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/payroll/cash-advance": {
      "post": {
        "tags": [
          "Cash Advance"
        ],
        "summary": "Request Cash Advance",
        "description": "Create a new cash advance request.\n\nStatus transitions:\n- pending (initial) \u2192 approved (by manager) \u2192 settled (at payroll calculation)\n- pending \u2192 rejected (by manager)\n\n**Business Logic:**\n- Pending CA cannot be deducted from payroll\n- Only approved CA is eligible for deduction\n- CA is automatically marked as settled when included in payroll calculation\n- settled_payroll_id tracks which payroll record settled this CA\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashAdvanceCreate"
              },
              "example": {
                "employee_id": 1,
                "amount": 2000.0,
                "request_date": "2026-05-20",
                "reason": "Medical emergency"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cash advance created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashAdvanceResponse"
                },
                "example": {
                  "id": 1,
                  "employee_id": 1,
                  "amount": 2000.0,
                  "request_date": "2026-05-20",
                  "reason": "Medical emergency",
                  "status": "pending",
                  "settled_payroll_id": null,
                  "created_at": "2026-05-20T14:30:00",
                  "updated_at": "2026-05-20T14:30:00"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Cash Advance"
        ],
        "summary": "List Cash Advances",
        "description": "Retrieve paginated list of cash advances with optional status filter",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Number of records to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 100
            },
            "description": "Maximum number of records to return"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "settled"
              ]
            },
            "description": "Filter by CA status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of cash advances",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CashAdvanceResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/payroll/cash-advance/{ca_id}": {
      "put": {
        "tags": [
          "Cash Advance"
        ],
        "summary": "Update Cash Advance Status",
        "description": "Update the status of a cash advance request.\n\nValid status values:\n- pending: Initial state\n- approved: Approved by manager\n- rejected: Rejected by manager\n- settled: Automatically set during payroll calculation\n\n**Note:** Usually only pending \u2192 approved or pending \u2192 rejected transitions are manual.\nsettled is set automatically by the payroll calculation system.\n",
        "parameters": [
          {
            "name": "ca_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Cash Advance ID"
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "settled"
              ]
            },
            "description": "New status"
          }
        ],
        "responses": {
          "200": {
            "description": "Status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashAdvanceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "\uc720\ud6a8\ud558\uc9c0 \uc54a\uc740 \uc0c1\ud0dc\uc785\ub2c8\ub2e4"
                }
              }
            }
          },
          "404": {
            "description": "Cash advance not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "CA\ub97c \ucc3e\uc744 \uc218 \uc5c6\uc2b5\ub2c8\ub2e4"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/payroll/attendance/{log_id}": {
      "put": {
        "tags": [
          "Attendance"
        ],
        "summary": "Update Attendance Record",
        "description": "Update an existing attendance record",
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Attendance Log ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendanceLogCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Attendance updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceLogResponse"
                }
              }
            }
          },
          "404": {
            "description": "Attendance record not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "\ucd9c\ud1f4\uadfc \uae30\ub85d\uc744 \ucc3e\uc744 \uc218 \uc5c6\uc2b5\ub2c8\ub2e4"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/payroll/holidays/{holiday_id}": {
      "delete": {
        "tags": [
          "Holidays"
        ],
        "summary": "Delete Philippine Holiday",
        "description": "Remove a registered Philippine holiday from the system",
        "parameters": [
          {
            "name": "holiday_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Holiday ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Holiday deleted successfully"
          },
          "404": {
            "description": "Holiday not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "\uacf5\ud734\uc77c\uc744 \ucc3e\uc744 \uc218 \uc5c6\uc2b5\ub2c8\ub2e4"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmployeeResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/EmployeeCreate"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "description": "Employee unique ID"
              },
              "is_active": {
                "type": "boolean",
                "description": "Active status"
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "description": "Record creation timestamp"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time",
                "description": "Last update timestamp"
              }
            }
          }
        ]
      },
...
    }
  }
}